Calculates threshold-performance data

This function is used by the threshperf_plot function. This is based on the `threshold_perf()` function from the `probably` package (a part of `tidymodels`) but expands the capability to allow for the range of metrics to include positive and negative predictive value.

threshperf(df, outcome, prediction, positive = "has_sepsis", thresholds = NULL)

Arguments

df
The df as a data.frame.

outcome
A character string containing the name of the column containing the outcomes (expressed as 0/1s).

prediction
A character string containing the name of the column containing the predictions.

Value

A data.frame containing the columns .threshold, .metric, .estimator, and .estimate

Examples

data(single_model_dataset)
threshperf(single_model_dataset, outcome = 'outcomes', prediction = 'predictions')

#>     .threshold .metric .estimator .estimate denom numer  PointEst          ll
#> 1         0.00    sens     binary 1.0000000   500   500 1.0000000 0.992375660
#> 2         0.01    sens     binary 0.9940000   500   497 0.9940000 0.982509748
#> 3         0.02    sens     binary 0.9840000   500   492 0.9840000 0.968748891
#> 4         0.03    sens     binary 0.9780000   500   489 0.9780000 0.961039139
#> 5         0.04    sens     binary 0.9700000   500   485 0.9700000 0.951096333
#> 6         0.05    sens     binary 0.9660000   500   483 0.9660000 0.946228634
#> 7         0.06    sens     binary 0.9640000   500   482 0.9640000 0.943815673
#> 8         0.07    sens     binary 0.9620000   500   481 0.9620000 0.941415227
#> 9         0.08    sens     binary 0.9600000   500   480 0.9600000 0.939026404
#> 10        0.09    sens     binary 0.9520000   500   476 0.9520000 0.929572803
#> 11        0.10    sens     binary 0.9380000   500   469 0.9380000 0.913340287
#> 12        0.11    sens     binary 0.9360000   500   468 0.9360000 0.911047591
#> 13        0.12    sens     binary 0.9320000   500   466 0.9320000 0.906479077
#> 14        0.13    sens     binary 0.9260000   500   463 0.9260000 0.899665247
#> 15        0.14    sens     binary 0.9180000   500   459 0.9180000 0.890645141
#> 16        0.15    sens     binary 0.9180000   500   459 0.9180000 0.890645141
#> 17        0.16    sens     binary 0.9120000   500   456 0.9120000 0.883923528
#> 18        0.17    sens     binary 0.9080000   500   454 0.9080000 0.879461262
#> 19        0.18    sens     binary 0.9060000   500   453 0.9060000 0.877235444
#> 20        0.19    sens     binary 0.9020000   500   451 0.9020000 0.872793949
#> 21        0.20    sens     binary 0.8980000   500   449 0.8980000 0.868365377
#> 22        0.21    sens     binary 0.8960000   500   448 0.8960000 0.866155729
#> 23        0.22    sens     binary 0.8880000   500   444 0.8880000 0.857346312
#> 24        0.23    sens     binary 0.8820000   500   441 0.8820000 0.850768012
#> 25        0.24    sens     binary 0.8780000   500   439 0.8780000 0.846395277
#> 26        0.25    sens     binary 0.8740000   500   437 0.8740000 0.842032282
#> 27        0.26    sens     binary 0.8720000   500   436 0.8720000 0.839854315
#> 28        0.27    sens     binary 0.8700000   500   435 0.8700000 0.837678638
#> 29        0.28    sens     binary 0.8680000   500   434 0.8680000 0.835505209
#> 30        0.29    sens     binary 0.8620000   500   431 0.8620000 0.828997986
#> 31        0.30    sens     binary 0.8560000   500   428 0.8560000 0.822509544
#> 32        0.31    sens     binary 0.8500000   500   425 0.8500000 0.816038919
#> 33        0.32    sens     binary 0.8440000   500   422 0.8440000 0.809585245
#> 34        0.33    sens     binary 0.8400000   500   420 0.8400000 0.805291821
#> 35        0.34    sens     binary 0.8380000   500   419 0.8380000 0.803147734
#> 36        0.35    sens     binary 0.8340000   500   417 0.8340000 0.798864684
#> 37        0.36    sens     binary 0.8260000   500   413 0.8260000 0.790318402
#> 38        0.37    sens     binary 0.8220000   500   411 0.8220000 0.786054811
#> 39        0.38    sens     binary 0.8160000   500   408 0.8160000 0.779670896
#> 40        0.39    sens     binary 0.8080000   500   404 0.8080000 0.771179625
#> 41        0.40    sens     binary 0.8000000   500   400 0.8000000 0.762710895
#> 42        0.41    sens     binary 0.8000000   500   400 0.8000000 0.762710895
#> 43        0.42    sens     binary 0.7820000   500   391 0.7820000 0.743733598
#> 44        0.43    sens     binary 0.7600000   500   380 0.7600000 0.720673203
#> 45        0.44    sens     binary 0.7260000   500   363 0.7260000 0.685294424
#> 46        0.45    sens     binary 0.7160000   500   358 0.7160000 0.674944057
#> 47        0.46    sens     binary 0.7060000   500   353 0.7060000 0.664617259
#> 48        0.47    sens     binary 0.6920000   500   346 0.6920000 0.650198001
#> 49        0.48    sens     binary 0.6760000   500   338 0.6760000 0.633771514
#> 50        0.49    sens     binary 0.6460000   500   323 0.6460000 0.603115998
#> 51        0.50    sens     binary 0.6360000   500   318 0.6360000 0.592937615
#> 52        0.51    sens     binary 0.6200000   500   310 0.6200000 0.576692489
#> 53        0.52    sens     binary 0.5920000   500   296 0.5920000 0.548379512
#> 54        0.53    sens     binary 0.5720000   500   286 0.5720000 0.528243850
#> 55        0.54    sens     binary 0.5580000   500   279 0.5580000 0.514191530
#> 56        0.55    sens     binary 0.5280000   500   264 0.5280000 0.484195772
#> 57        0.56    sens     binary 0.5060000   500   253 0.5060000 0.462298639
#> 58        0.57    sens     binary 0.4940000   500   247 0.4940000 0.450390131
#> 59        0.58    sens     binary 0.4740000   500   237 0.4740000 0.430598114
#> 60        0.59    sens     binary 0.4640000   500   232 0.4640000 0.420728187
#> 61        0.60    sens     binary 0.4540000   500   227 0.4540000 0.410875727
#> 62        0.61    sens     binary 0.4380000   500   219 0.4380000 0.395148344
#> 63        0.62    sens     binary 0.4300000   500   215 0.4300000 0.387301648
#> 64        0.63    sens     binary 0.4200000   500   210 0.4200000 0.377509351
#> 65        0.64    sens     binary 0.4040000   500   202 0.4040000 0.361879225
#> 66        0.65    sens     binary 0.3920000   500   196 0.3920000 0.350187378
#> 67        0.66    sens     binary 0.3780000   500   189 0.3780000 0.336580794
#> 68        0.67    sens     binary 0.3680000   500   184 0.3680000 0.326884548
#> 69        0.68    sens     binary 0.3580000   500   179 0.3580000 0.317207603
#> 70        0.69    sens     binary 0.3500000   500   175 0.3500000 0.309480173
#> 71        0.70    sens     binary 0.3400000   500   170 0.3400000 0.299838858
#> 72        0.71    sens     binary 0.3280000   500   164 0.3280000 0.288296190
#> 73        0.72    sens     binary 0.3180000   500   159 0.3180000 0.278700261
#> 74        0.73    sens     binary 0.3120000   500   156 0.3120000 0.272952949
#> 75        0.74    sens     binary 0.3020000   500   151 0.3020000 0.263391551
#> 76        0.75    sens     binary 0.2840000   500   142 0.2840000 0.246237772
#> 77        0.76    sens     binary 0.2820000   500   141 0.2820000 0.244336433
#> 78        0.77    sens     binary 0.2720000   500   136 0.2720000 0.234844092
#> 79        0.78    sens     binary 0.2620000   500   131 0.2620000 0.225376269
#> 80        0.79    sens     binary 0.2500000   500   125 0.2500000 0.214048497
#> 81        0.80    sens     binary 0.2400000   500   120 0.2400000 0.204637860
#> 82        0.81    sens     binary 0.2260000   500   113 0.2260000 0.191509781
#> 83        0.82    sens     binary 0.2080000   500   104 0.2080000 0.174716312
#> 84        0.83    sens     binary 0.2000000   500   100 0.2000000 0.167285499
#> 85        0.84    sens     binary 0.1900000   500    95 0.1900000 0.158027379
#> 86        0.85    sens     binary 0.1780000   500    89 0.1780000 0.146964887
#> 87        0.86    sens     binary 0.1580000   500    79 0.1580000 0.128652698
#> 88        0.87    sens     binary 0.1460000   500    73 0.1460000 0.117748767
#> 89        0.88    sens     binary 0.1380000   500    69 0.1380000 0.110518009
#> 90        0.89    sens     binary 0.1340000   500    67 0.1340000 0.106915002
#> 91        0.90    sens     binary 0.1240000   500    62 0.1240000 0.097946091
#> 92        0.91    sens     binary 0.1180000   500    59 0.1180000 0.092593008
#> 93        0.92    sens     binary 0.1040000   500    52 0.1040000 0.080194207
#> 94        0.93    sens     binary 0.0960000   500    48 0.0960000 0.073173505
#> 95        0.94    sens     binary 0.0720000   500    36 0.0720000 0.052458031
#> 96        0.95    sens     binary 0.0660000   500    33 0.0660000 0.047378526
#> 97        0.96    sens     binary 0.0540000   500    27 0.0540000 0.037374350
#> 98        0.97    sens     binary 0.0360000   500    18 0.0360000 0.022891061
#> 99        0.98    sens     binary 0.0280000   500    14 0.0280000 0.016750975
#> 100       0.99    sens     binary 0.0200000   500    10 0.0200000 0.010899184
#> 101       1.00    sens     binary 0.0120000   500     6 0.0120000 0.005510995
#> 102       0.00    spec     binary 0.0000000   500     0 0.0000000 0.000000000
#> 103       0.01    spec     binary 0.0080000   500     4 0.0080000 0.003115315
#> 104       0.02    spec     binary 0.0180000   500     9 0.0180000 0.009498193
#> 105       0.03    spec     binary 0.0280000   500    14 0.0280000 0.016750975
#> 106       0.04    spec     binary 0.0360000   500    18 0.0360000 0.022891061
#> 107       0.05    spec     binary 0.0540000   500    27 0.0540000 0.037374350
#> 108       0.06    spec     binary 0.0660000   500    33 0.0660000 0.047378526
#> 109       0.07    spec     binary 0.0840000   500    42 0.0840000 0.062744137
#> 110       0.08    spec     binary 0.0940000   500    47 0.0940000 0.071426409
#> 111       0.09    spec     binary 0.1040000   500    52 0.1040000 0.080194207
#> 112       0.10    spec     binary 0.1160000   500    58 0.1160000 0.090813657
#> 113       0.11    spec     binary 0.1420000   500    71 0.1420000 0.114129360
#> 114       0.12    spec     binary 0.1440000   500    72 0.1440000 0.115938074
#> 115       0.13    spec     binary 0.1540000   500    77 0.1540000 0.125010680
#> 116       0.14    spec     binary 0.1620000   500    81 0.1620000 0.132301788
#> 117       0.15    spec     binary 0.1740000   500    87 0.1740000 0.143289472
#> 118       0.16    spec     binary 0.1900000   500    95 0.1900000 0.158027379
#> 119       0.17    spec     binary 0.2000000   500   100 0.2000000 0.167285499
#> 120       0.18    spec     binary 0.2140000   500   107 0.2140000 0.180303000
#> 121       0.19    spec     binary 0.2240000   500   112 0.2240000 0.189638971
#> 122       0.20    spec     binary 0.2460000   500   123 0.2460000 0.210280992
#> 123       0.21    spec     binary 0.2560000   500   128 0.2560000 0.219707708
#> 124       0.22    spec     binary 0.2660000   500   133 0.2660000 0.229160402
#> 125       0.23    spec     binary 0.2780000   500   139 0.2780000 0.240536608
#> 126       0.24    spec     binary 0.2880000   500   144 0.2880000 0.250043271
#> 127       0.25    spec     binary 0.3000000   500   150 0.3000000 0.261481932
#> 128       0.26    spec     binary 0.3160000   500   158 0.3160000 0.276783628
#> 129       0.27    spec     binary 0.3300000   500   165 0.3300000 0.290217900
#> 130       0.28    spec     binary 0.3400000   500   170 0.3400000 0.299838858
#> 131       0.29    spec     binary 0.3500000   500   175 0.3500000 0.309480173
#> 132       0.30    spec     binary 0.3520000   500   176 0.3520000 0.311410842
#> 133       0.31    spec     binary 0.3640000   500   182 0.3640000 0.323011436
#> 134       0.32    spec     binary 0.3820000   500   191 0.3820000 0.340464625
#> 135       0.33    spec     binary 0.4020000   500   201 0.4020000 0.359928740
#> 136       0.34    spec     binary 0.4080000   500   204 0.4080000 0.365782391
#> 137       0.35    spec     binary 0.4140000   500   207 0.4140000 0.371642609
#> 138       0.36    spec     binary 0.4200000   500   210 0.4200000 0.377509351
#> 139       0.37    spec     binary 0.4320000   500   216 0.4320000 0.389262254
#> 140       0.38    spec     binary 0.4420000   500   221 0.4420000 0.399075953
#> 141       0.39    spec     binary 0.4520000   500   226 0.4520000 0.408907338
#> 142       0.40    spec     binary 0.4680000   500   234 0.4680000 0.424674066
#> 143       0.41    spec     binary 0.4840000   500   242 0.4840000 0.440485443
#> 144       0.42    spec     binary 0.5000000   500   250 0.5000000 0.456341265
#> 145       0.43    spec     binary 0.5080000   500   254 0.5080000 0.464285817
#> 146       0.44    spec     binary 0.5120000   500   256 0.5120000 0.468262253
#> 147       0.45    spec     binary 0.5360000   500   268 0.5360000 0.492179234
#> 148       0.46    spec     binary 0.5560000   500   278 0.5560000 0.512186893
#> 149       0.47    spec     binary 0.5740000   500   287 0.5740000 0.530254176
#> 150       0.48    spec     binary 0.5880000   500   294 0.5880000 0.544346592
#> 151       0.49    spec     binary 0.6040000   500   302 0.6040000 0.560495841
#> 152       0.50    spec     binary 0.6100000   500   305 0.6100000 0.566563975
#> 153       0.51    spec     binary 0.6260000   500   313 0.6260000 0.582778666
#> 154       0.52    spec     binary 0.6400000   500   320 0.6400000 0.597006617
#> 155       0.53    spec     binary 0.6520000   500   326 0.6520000 0.609232503
#> 156       0.54    spec     binary 0.6680000   500   334 0.6680000 0.625578663
#> 157       0.55    spec     binary 0.6980000   500   349 0.6980000 0.656372312
#> 158       0.56    spec     binary 0.7220000   500   361 0.7220000 0.681151400
#> 159       0.57    spec     binary 0.7420000   500   371 0.7420000 0.701906019
#> 160       0.58    spec     binary 0.7540000   500   377 0.7540000 0.714407827
#> 161       0.59    spec     binary 0.7660000   500   383 0.7660000 0.726948549
#> 162       0.60    spec     binary 0.7760000   500   388 0.7760000 0.737430335
#> 163       0.61    spec     binary 0.7900000   500   395 0.7900000 0.752155153
#> 164       0.62    spec     binary 0.8000000   500   400 0.8000000 0.762710895
#> 165       0.63    spec     binary 0.8020000   500   401 0.8020000 0.764826021
#> 166       0.64    spec     binary 0.8040000   500   402 0.8040000 0.766942508
#> 167       0.65    spec     binary 0.8100000   500   405 0.8100000 0.773300288
#> 168       0.66    spec     binary 0.8180000   500   409 0.8180000 0.781797365
#> 169       0.67    spec     binary 0.8180000   500   409 0.8180000 0.781797365
#> 170       0.68    spec     binary 0.8200000   500   410 0.8200000 0.783925330
#> 171       0.69    spec     binary 0.8200000   500   410 0.8200000 0.783925330
#> 172       0.70    spec     binary 0.8220000   500   411 0.8220000 0.786054811
#> 173       0.71    spec     binary 0.8280000   500   414 0.8280000 0.792452553
#> 174       0.72    spec     binary 0.8340000   500   417 0.8340000 0.798864684
#> 175       0.73    spec     binary 0.8360000   500   418 0.8360000 0.801005364
#> 176       0.74    spec     binary 0.8480000   500   424 0.8480000 0.813885852
#> 177       0.75    spec     binary 0.8520000   500   426 0.8520000 0.818193870
#> 178       0.76    spec     binary 0.8560000   500   428 0.8560000 0.822509544
#> 179       0.77    spec     binary 0.8560000   500   428 0.8560000 0.822509544
#> 180       0.78    spec     binary 0.8560000   500   428 0.8560000 0.822509544
#> 181       0.79    spec     binary 0.8620000   500   431 0.8620000 0.828997986
#> 182       0.80    spec     binary 0.8680000   500   434 0.8680000 0.835505209
#> 183       0.81    spec     binary 0.8820000   500   441 0.8820000 0.850768012
#> 184       0.82    spec     binary 0.8880000   500   444 0.8880000 0.857346312
#> 185       0.83    spec     binary 0.8980000   500   449 0.8980000 0.868365377
#> 186       0.84    spec     binary 0.9000000   500   450 0.9000000 0.870578090
#> 187       0.85    spec     binary 0.9080000   500   454 0.9080000 0.879461262
#> 188       0.86    spec     binary 0.9120000   500   456 0.9120000 0.883923528
#> 189       0.87    spec     binary 0.9200000   500   460 0.9200000 0.892893683
#> 190       0.88    spec     binary 0.9240000   500   462 0.9240000 0.897403586
#> 191       0.89    spec     binary 0.9280000   500   464 0.9280000 0.901931596
#> 192       0.90    spec     binary 0.9320000   500   466 0.9320000 0.906479077
#> 193       0.91    spec     binary 0.9400000   500   470 0.9400000 0.915638937
#> 194       0.92    spec     binary 0.9480000   500   474 0.9480000 0.924898856
#> 195       0.93    spec     binary 0.9560000   500   478 0.9560000 0.934280557
#> 196       0.94    spec     binary 0.9620000   500   481 0.9620000 0.941415227
#> 197       0.95    spec     binary 0.9680000   500   484 0.9680000 0.948655131
#> 198       0.96    spec     binary 0.9720000   500   486 0.9720000 0.953553597
#> 199       0.97    spec     binary 0.9780000   500   489 0.9780000 0.961039139
#> 200       0.98    spec     binary 0.9820000   500   491 0.9820000 0.966148329
#> 201       0.99    spec     binary 0.9920000   500   496 0.9920000 0.979612964
#> 202       1.00    spec     binary 0.9960000   500   498 0.9960000 0.985534285
#> 203       0.00     ppv     binary 0.5000000  1000   500 0.5000000 0.469069600
#> 204       0.01     ppv     binary 0.5005035   993   497 0.5005035 0.469462791
#> 205       0.02     ppv     binary 0.5005086   983   492 0.5005086 0.469311007
#> 206       0.03     ppv     binary 0.5015385   975   489 0.5015385 0.470209675
#> 207       0.04     ppv     binary 0.5015512   967   485 0.5015512 0.470093516
#> 208       0.05     ppv     binary 0.5052301   956   483 0.5052301 0.473579522
#> 209       0.06     ppv     binary 0.5079031   949   482 0.5079031 0.476127773
#> 210       0.07     ppv     binary 0.5122471   939   481 0.5122471 0.480291422
#> 211       0.08     ppv     binary 0.5144695   933   480 0.5144695 0.482406152
#> 212       0.09     ppv     binary 0.5151515   924   476 0.5151515 0.482931265
#> 213       0.10     ppv     binary 0.5148189   911   469 0.5148189 0.482370816
#> 214       0.11     ppv     binary 0.5217391   897   468 0.5217391 0.489026366
#> 215       0.12     ppv     binary 0.5212528   894   466 0.5212528 0.488485987
#> 216       0.13     ppv     binary 0.5225734   886   463 0.5225734 0.489657269
#> 217       0.14     ppv     binary 0.5227790   878   459 0.5227790 0.489713253
#> 218       0.15     ppv     binary 0.5263761   872   459 0.5263761 0.493192845
#> 219       0.16     ppv     binary 0.5296167   861   456 0.5296167 0.496219982
#> 220       0.17     ppv     binary 0.5316159   854   454 0.5316159 0.498081878
#> 221       0.18     ppv     binary 0.5354610   846   453 0.5354610 0.501768729
#> 222       0.19     ppv     binary 0.5375447   839   451 0.5375447 0.503712893
#> 223       0.20     ppv     binary 0.5435835   826   449 0.5435835 0.509491745
#> 224       0.21     ppv     binary 0.5463415   820   448 0.5463415 0.512129079
#> 225       0.22     ppv     binary 0.5474723   811   444 0.5474723 0.513072197
#> 226       0.23     ppv     binary 0.5498753   802   441 0.5498753 0.515287122
#> 227       0.24     ppv     binary 0.5522013   795   439 0.5522013 0.517466091
#> 228       0.25     ppv     binary 0.5552732   787   437 0.5552732 0.520369604
#> 229       0.26     ppv     binary 0.5604113   778   436 0.5604113 0.525322341
#> 230       0.27     ppv     binary 0.5649351   770   435 0.5649351 0.529681246
#> 231       0.28     ppv     binary 0.5680628   764   434 0.5680628 0.532684153
#> 232       0.29     ppv     binary 0.5701058   756   431 0.5701058 0.534549409
#> 233       0.30     ppv     binary 0.5691489   752   428 0.5691489 0.533492936
#> 234       0.31     ppv     binary 0.5720054   743   425 0.5720054 0.536147407
#> 235       0.32     ppv     binary 0.5772914   731   422 0.5772914 0.541168510
#> 236       0.33     ppv     binary 0.5841446   719   420 0.5841446 0.547764628
#> 237       0.34     ppv     binary 0.5860140   715   419 0.5860140 0.549545108
#> 238       0.35     ppv     binary 0.5873239   710   417 0.5873239 0.550735706
#> 239       0.36     ppv     binary 0.5874822   703   413 0.5874822 0.550712095
#> 240       0.37     ppv     binary 0.5913669   695   411 0.5913669 0.554414862
#> 241       0.38     ppv     binary 0.5938865   687   408 0.5938865 0.556739332
#> 242       0.39     ppv     binary 0.5958702   678   404 0.5958702 0.558492647
#> 243       0.40     ppv     binary 0.6006006   666   400 0.6006006 0.562929025
#> 244       0.41     ppv     binary 0.6079027   658   400 0.6079027 0.570076132
#> 245       0.42     ppv     binary 0.6099844   641   391 0.6099844 0.571677226
#> 246       0.43     ppv     binary 0.6070288   626   380 0.6070288 0.568227159
#> 247       0.44     ppv     binary 0.5980231   607   363 0.5980231 0.558520198
#> 248       0.45     ppv     binary 0.6067797   590   358 0.6067797 0.566796080
#> 249       0.46     ppv     binary 0.6139130   575   353 0.6139130 0.573488804
#> 250       0.47     ppv     binary 0.6189624   559   346 0.6189624 0.578021346
#> 251       0.48     ppv     binary 0.6213235   544   338 0.6213235 0.579846354
#> 252       0.49     ppv     binary 0.6199616   521   323 0.6199616 0.577547337
#> 253       0.50     ppv     binary 0.6198830   513   318 0.6198830 0.577133726
#> 254       0.51     ppv     binary 0.6237425   497   310 0.6237425 0.580355647
#> 255       0.52     ppv     binary 0.6218487   476   296 0.6218487 0.577473820
#> 256       0.53     ppv     binary 0.6217391   460   286 0.6217391 0.576586426
#> 257       0.54     ppv     binary 0.6269663   445   279 0.6269663 0.581126261
#> 258       0.55     ppv     binary 0.6361446   415   264 0.6361446 0.588803967
#> 259       0.56     ppv     binary 0.6454082   392   253 0.6454082 0.596849004
#> 260       0.57     ppv     binary 0.6569149   376   247 0.6569149 0.607559399
#> 261       0.58     ppv     binary 0.6583333   360   237 0.6583333 0.607900753
#> 262       0.59     ppv     binary 0.6647564   349   232 0.6647564 0.613672853
#> 263       0.60     ppv     binary 0.6696165   339   227 0.6696165 0.617891862
#> 264       0.61     ppv     binary 0.6759259   324   219 0.6759259 0.623159962
#> 265       0.62     ppv     binary 0.6825397   315   215 0.6825397 0.629199164
#> 266       0.63     ppv     binary 0.6796117   309   210 0.6796117 0.625651501
#> 267       0.64     ppv     binary 0.6733333   300   202 0.6733333 0.618362209
#> 268       0.65     ppv     binary 0.6735395   291   196 0.6735395 0.617706496
#> 269       0.66     ppv     binary 0.6750000   280   189 0.6750000 0.618091772
#> 270       0.67     ppv     binary 0.6690909   275   184 0.6690909 0.611483431
#> 271       0.68     ppv     binary 0.6654275   269   179 0.6654275 0.607062799
#> 272       0.69     ppv     binary 0.6603774   265   175 0.6603774 0.601429213
#> 273       0.70     ppv     binary 0.6563707   259   170 0.6563707 0.596625411
#> 274       0.71     ppv     binary 0.6560000   250   164 0.6560000 0.595153188
#> 275       0.72     ppv     binary 0.6570248   242   159 0.6570248 0.595181108
#> 276       0.73     ppv     binary 0.6554622   238   156 0.6554622 0.593049056
#> 277       0.74     ppv     binary 0.6651982   227   151 0.6651982 0.601509085
#> 278       0.75     ppv     binary 0.6574074   216   142 0.6574074 0.591863131
#> 279       0.76     ppv     binary 0.6619718   213   141 0.6619718 0.596075845
#> 280       0.77     ppv     binary 0.6538462   208   136 0.6538462 0.586931510
#> 281       0.78     ppv     binary 0.6453202   203   131 0.6453202 0.577367342
#> 282       0.79     ppv     binary 0.6443299   194   125 0.6443299 0.574762260
#> 283       0.80     ppv     binary 0.6451613   186   120 0.6451613 0.574098976
#> 284       0.81     ppv     binary 0.6569767   172   113 0.6569767 0.583298054
#> 285       0.82     ppv     binary 0.6500000   160   104 0.6500000 0.573364139
#> 286       0.83     ppv     binary 0.6622517   151   100 0.6622517 0.583625171
#> 287       0.84     ppv     binary 0.6551724   145    95 0.6551724 0.574702783
#> 288       0.85     ppv     binary 0.6592593   135    89 0.6592593 0.575893096
#> 289       0.86     ppv     binary 0.6422764   123    79 0.6422764 0.554439738
#> 290       0.87     ppv     binary 0.6460177   113    73 0.6460177 0.554375519
#> 291       0.88     ppv     binary 0.6448598   107    69 0.6448598 0.550607908
#> 292       0.89     ppv     binary 0.6504854   103    67 0.6504854 0.554500280
#> 293       0.90     ppv     binary 0.6458333    96    62 0.6458333 0.546243023
#> 294       0.91     ppv     binary 0.6629213    89    59 0.6629213 0.559788804
#> 295       0.92     ppv     binary 0.6666667    78    52 0.6666667 0.556414097
#> 296       0.93     ppv     binary 0.6857143    70    48 0.6857143 0.569728620
#> 297       0.94     ppv     binary 0.6545455    55    36 0.6545455 0.522539035
#> 298       0.95     ppv     binary 0.6734694    49    33 0.6734694 0.533792053
#> 299       0.96     ppv     binary 0.6585366    41    27 0.6585366 0.505498365
#> 300       0.97     ppv     binary 0.6206897    29    18 0.6206897 0.440025442
#> 301       0.98     ppv     binary 0.6086957    23    14 0.6086957 0.407855224
#> 302       0.99     ppv     binary 0.7142857    14    10 0.7142857 0.453509157
#> 303       1.00     ppv     binary 0.7500000     8     6 0.7500000 0.409275430
#> 304       0.01     npv     binary 0.5714286     7     4 0.5714286 0.250458365
#> 305       0.02     npv     binary 0.5294118    17     9 0.5294118 0.309632351
#> 306       0.03     npv     binary 0.5600000    25    14 0.5600000 0.370673319
#> 307       0.04     npv     binary 0.5454545    33    18 0.5454545 0.379859385
#> 308       0.05     npv     binary 0.6136364    44    27 0.6136364 0.466235754
#> 309       0.06     npv     binary 0.6470588    51    33 0.6470588 0.509860585
#> 310       0.07     npv     binary 0.6885246    61    42 0.6885246 0.564085656
#> 311       0.08     npv     binary 0.7014925    67    47 0.7014925 0.583447863
#> 312       0.09     npv     binary 0.6842105    76    52 0.6842105 0.573003537
#> 313       0.10     npv     binary 0.6516854    89    58 0.6516854 0.548293180
#> 314       0.11     npv     binary 0.6893204   103    71 0.6893204 0.594500202
#> 315       0.12     npv     binary 0.6792453   106    72 0.6792453 0.585461719
#> 316       0.13     npv     binary 0.6754386   114    77 0.6754386 0.584990573
#> 317       0.14     npv     binary 0.6639344   122    81 0.6639344 0.576248747
#> 318       0.15     npv     binary 0.6796875   128    87 0.6796875 0.594634112
#> 319       0.16     npv     binary 0.6834532   139    95 0.6834532 0.602083090
#> 320       0.17     npv     binary 0.6849315   146   100 0.6849315 0.605659165
#> 321       0.18     npv     binary 0.6948052   154   107 0.6948052 0.618069293
#> 322       0.19     npv     binary 0.6956522   161   112 0.6956522 0.620703013
#> 323       0.20     npv     binary 0.7068966   174   123 0.7068966 0.635379327
#> 324       0.21     npv     binary 0.7111111   180   128 0.7111111 0.641033521
#> 325       0.22     npv     binary 0.7037037   189   133 0.7037037 0.635070768
#> 326       0.23     npv     binary 0.7020202   198   139 0.7020202 0.634960956
#> 327       0.24     npv     binary 0.7024390   205   144 0.7024390 0.636597862
#> 328       0.25     npv     binary 0.7042254   213   150 0.7042254 0.639754371
#> 329       0.26     npv     binary 0.7117117   222   158 0.7117117 0.648924839
#> 330       0.27     npv     binary 0.7173913   230   165 0.7173913 0.655998752
#> 331       0.28     npv     binary 0.7203390   236   170 0.7203390 0.659897481
#> 332       0.29     npv     binary 0.7172131   244   175 0.7172131 0.657677421
#> 333       0.30     npv     binary 0.7096774   248   176 0.7096774 0.650327654
#> 334       0.31     npv     binary 0.7081712   257   182 0.7081712 0.649851584
#> 335       0.32     npv     binary 0.7100372   269   191 0.7100372 0.653158868
#> 336       0.33     npv     binary 0.7153025   281   201 0.7153025 0.659912215
#> 337       0.34     npv     binary 0.7157895   285   204 0.7157895 0.660825166
#> 338       0.35     npv     binary 0.7137931   290   207 0.7137931 0.659243174
#> 339       0.36     npv     binary 0.7070707   297   210 0.7070707 0.652931579
#> 340       0.37     npv     binary 0.7081967   305   216 0.7081967 0.654841675
#> 341       0.38     npv     binary 0.7060703   313   221 0.7060703 0.653347897
#> 342       0.39     npv     binary 0.7018634   322   226 0.7018634 0.649758282
#> 343       0.40     npv     binary 0.7005988   334   234 0.7005988 0.649427177
#> 344       0.41     npv     binary 0.7076023   342   242 0.7076023 0.657301714
#> 345       0.42     npv     binary 0.6963788   359   250 0.6963788 0.646941300
#> 346       0.43     npv     binary 0.6791444   374   254 0.6791444 0.630219454
#> 347       0.44     npv     binary 0.6513995   393   256 0.6513995 0.603026747
#> 348       0.45     npv     binary 0.6536585   410   268 0.6536585 0.606368528
#> 349       0.46     npv     binary 0.6541176   425   278 0.6541176 0.607697350
#> 350       0.47     npv     binary 0.6507937   441   287 0.6507937 0.605171873
#> 351       0.48     npv     binary 0.6447368   456   294 0.6447368 0.599767840
#> 352       0.49     npv     binary 0.6304802   479   302 0.6304802 0.586376860
#> 353       0.50     npv     binary 0.6262834   487   305 0.6262834 0.582484580
#> 354       0.51     npv     binary 0.6222664   503   313 0.6222664 0.579121719
#> 355       0.52     npv     binary 0.6106870   524   320 0.6106870 0.568277339
#> 356       0.53     npv     binary 0.6037037   540   326 0.6037037 0.561855949
#> 357       0.54     npv     binary 0.6018018   555   334 0.6018018 0.560509621
#> 358       0.55     npv     binary 0.5965812   585   349 0.5965812 0.556321802
#> 359       0.56     npv     binary 0.5937500   608   361 0.5937500 0.554241032
#> 360       0.57     npv     binary 0.5945513   624   371 0.5945513 0.555563593
#> 361       0.58     npv     binary 0.5890625   640   377 0.5890625 0.550523581
#> 362       0.59     npv     binary 0.5883257   651   383 0.5883257 0.550110510
#> 363       0.60     npv     binary 0.5869894   661   388 0.5869894 0.549056431
#> 364       0.61     npv     binary 0.5843195   676   395 0.5843195 0.546793335
#> 365       0.62     npv     binary 0.5839416   685   400 0.5839416 0.546661827
#> 366       0.63     npv     binary 0.5803184   691   401 0.5803184 0.543177369
#> 367       0.64     npv     binary 0.5742857   700   402 0.5742857 0.537349360
#> 368       0.65     npv     binary 0.5712271   709   405 0.5712271 0.534510856
#> 369       0.66     npv     binary 0.5680556   720   409 0.5680556 0.531606826
#> 370       0.67     npv     binary 0.5641379   725   409 0.5641379 0.527798607
#> 371       0.68     npv     binary 0.5608755   731   410 0.5608755 0.524673729
#> 372       0.69     npv     binary 0.5578231   735   410 0.5578231 0.521710039
#> 373       0.70     npv     binary 0.5546559   741   411 0.5546559 0.518680467
#> 374       0.71     npv     binary 0.5520000   750   414 0.5520000 0.516234986
#> 375       0.72     npv     binary 0.5501319   758   417 0.5501319 0.514552456
#> 376       0.73     npv     binary 0.5485564   762   418 0.5485564 0.513067555
#> 377       0.74     npv     binary 0.5485123   773   424 0.5485123 0.513277217
#> 378       0.75     npv     binary 0.5433673   784   426 0.5433673 0.508372898
#> 379       0.76     npv     binary 0.5438374   787   428 0.5438374 0.508910319
#> 380       0.77     npv     binary 0.5404040   792   428 0.5404040 0.505584068
#> 381       0.78     npv     binary 0.5370138   797   428 0.5370138 0.502301426
#> 382       0.79     npv     binary 0.5347395   806   431 0.5347395 0.500221064
#> 383       0.80     npv     binary 0.5331695   814   434 0.5331695 0.498821278
#> 384       0.81     npv     binary 0.5326087   828   441 0.5326087 0.498552110
#> 385       0.82     npv     binary 0.5285714   840   444 0.5285714 0.494760694
#> 386       0.83     npv     binary 0.5288575   849   449 0.5288575 0.495226118
#> 387       0.84     npv     binary 0.5263158   855   450 0.5263158 0.492804583
#> 388       0.85     npv     binary 0.5248555   865   454 0.5248555 0.491539874
#> 389       0.86     npv     binary 0.5199544   877   456 0.5199544 0.486874154
#> 390       0.87     npv     binary 0.5186020   887   460 0.5186020 0.485710894
#> 391       0.88     npv     binary 0.5173572   893   462 0.5173572 0.484578977
#> 392       0.89     npv     binary 0.5172798   897   464 0.5172798 0.484574750
#> 393       0.90     npv     binary 0.5154867   904   466 0.5154867 0.482912050
#> 394       0.91     npv     binary 0.5159166   911   470 0.5159166 0.483466080
#> 395       0.92     npv     binary 0.5140998   922   474 0.5140998 0.481847072
#> 396       0.93     npv     binary 0.5139785   930   478 0.5139785 0.481864764
#> 397       0.94     npv     binary 0.5089947   945   481 0.5089947 0.477149201
#> 398       0.95     npv     binary 0.5089380   951   484 0.5089380 0.477192943
#> 399       0.96     npv     binary 0.5067779   959   486 0.5067779 0.475171664
#> 400       0.97     npv     binary 0.5036045   971   489 0.5036045 0.472204046
#> 401       0.98     npv     binary 0.5025589   977   491 0.5025589 0.471258294
#> 402       0.99     npv     binary 0.5030426   986   496 0.5030426 0.471882996
#> 403       1.00     npv     binary 0.5020161   992   498 0.5020161 0.470954214
#>             ul
#> 1   1.00000000
#> 2   0.99795740
#> 3   0.99187075
#> 4   0.98767199
#> 5   0.98173679
#> 6   0.97866548
#> 7   0.97710894
#> 8   0.97553988
#> 9   0.97395920
#> 10  0.96753479
#> 11  0.95598079
#> 12  0.95430398
#> 13  0.95093349
#> 14  0.94583882
#> 15  0.93898091
#> 16  0.93898091
#> 17  0.93379402
#> 18  0.93031728
#> 19  0.92857359
#> 20  0.92507608
#> 21  0.92156565
#> 22  0.91980579
#> 23  0.91273720
#> 24  0.90740699
#> 25  0.90384072
#> 26  0.90026471
#> 27  0.89847318
#> 28  0.89667935
#> 29  0.89488328
#> 30  0.88948199
#> 31  0.88406193
#> 32  0.87862404
#> 33  0.87316921
#> 34  0.86952363
#> 35  0.86769821
#> 36  0.86404226
#> 37  0.85671053
#> 38  0.85303511
#> 39  0.84751052
#> 40  0.84012378
#> 41  0.83271450
#> 42  0.83271450
#> 43  0.81596627
#> 44  0.79536214
#> 45  0.76325937
#> 46  0.75376223
#> 47  0.74424151
#> 48  0.73087425
#> 49  0.71554472
#> 50  0.68665769
#> 51  0.67698856
#> 52  0.66147767
#> 53  0.63421761
#> 54  0.61465824
#> 55  0.60092405
#> 56  0.57137727
#> 57  0.54960987
#> 58  0.53770136
#> 59  0.51779835
#> 60  0.50782077
#> 61  0.49782571
#> 62  0.48179707
#> 63  0.47376576
#> 64  0.46371054
#> 65  0.44758465
#> 66  0.43545948
#> 67  0.42127955
#> 68  0.41112828
#> 69  0.40095771
#> 70  0.39280713
#> 71  0.38260093
#> 72  0.37032658
#> 73  0.36007500
#> 74  0.35391380
#> 75  0.34362769
#> 76  0.32505594
#> 77  0.32298778
#> 78  0.31263261
#> 79  0.30225292
#> 80  0.28976367
#> 81  0.27932680
#> 82  0.26466836
#> 83  0.24573630
#> 84  0.23728911
#> 85  0.22669971
#> 86  0.21394519
#> 87  0.19256235
#> 88  0.17964927
#> 89  0.17100201
#> 90  0.16666602
#> 91  0.15578741
#> 92  0.14923199
#> 93  0.13384427
#> 94  0.12498696
#> 95  0.09806840
#> 96  0.09123940
#> 97  0.07742656
#> 98  0.05618433
#> 99  0.04644640
#> 100 0.03642018
#> 101 0.02593036
#> 102 0.00762434
#> 103 0.02038704
#> 104 0.03385167
#> 105 0.04644640
#> 106 0.05618433
#> 107 0.07742656
#> 108 0.09123940
#> 109 0.11159931
#> 110 0.12276456
#> 111 0.13384427
#> 112 0.14704184
#> 113 0.17532967
#> 114 0.17749046
#> 115 0.18826536
#> 116 0.19685227
#> 117 0.20968160
#> 118 0.22669971
#> 119 0.23728911
#> 120 0.25205812
#> 121 0.26256966
#> 122 0.28559217
#> 123 0.29601297
#> 124 0.30640779
#> 125 0.31884860
#> 126 0.32918945
#> 127 0.34156780
#> 128 0.35802213
#> 129 0.37237438
#> 130 0.38260093
#> 131 0.39280713
#> 132 0.39484596
#> 133 0.40706238
#> 134 0.42533472
#> 135 0.44556563
#> 136 0.45162049
#> 137 0.45766878
#> 138 0.46371054
#> 139 0.47577466
#> 140 0.48580847
#> 141 0.49582460
#> 142 0.51181389
#> 143 0.52775854
#> 144 0.54365873
#> 145 0.55159219
#> 146 0.55555476
#> 147 0.57927181
#> 148 0.59895918
#> 149 0.61661742
#> 150 0.63031152
#> 151 0.64591830
#> 152 0.65175867
#> 153 0.66730000
#> 154 0.68085857
#> 155 0.69244970
#> 156 0.70785956
#> 157 0.73660845
#> 158 0.75946339
#> 159 0.77840380
#> 160 0.78971901
#> 161 0.80099530
#> 162 0.81036103
#> 163 0.82342273
#> 164 0.83271450
#> 165 0.83456888
#> 166 0.83642189
#> 167 0.84197262
#> 168 0.84935355
#> 169 0.84935355
#> 170 0.85119509
#> 171 0.85119509
#> 172 0.85303511
#> 173 0.85854588
#> 174 0.86404226
#> 175 0.86587108
#> 176 0.87680761
#> 177 0.88043859
#> 178 0.88406193
#> 179 0.88406193
#> 180 0.88406193
#> 181 0.88948199
#> 182 0.89488328
#> 183 0.90740699
#> 184 0.91273720
#> 185 0.92156565
#> 186 0.92332244
#> 187 0.93031728
#> 188 0.93379402
#> 189 0.94070187
#> 190 0.94413097
#> 191 0.94754197
#> 192 0.95093349
#> 193 0.95765164
#> 194 0.96426973
#> 195 0.97076604
#> 196 0.97553988
#> 197 0.98020849
#> 198 0.98324903
#> 199 0.98767199
#> 200 0.99050181
#> 201 0.99688468
#> 202 0.99890237
#> 203 0.53093040
#> 204 0.53154038
#> 205 0.53170233
#> 206 0.53285517
#> 207 0.53299659
#> 208 0.53683886
#> 209 0.53961462
#> 210 0.54410292
#> 211 0.54641409
#> 212 0.54724630
#> 213 0.54714249
#> 214 0.55426649
#> 215 0.55383774
#> 216 0.55529456
#> 217 0.55564637
#> 218 0.55932808
#> 219 0.56275036
#> 220 0.56486682
#> 221 0.56883268
#> 222 0.57103426
#> 223 0.57727182
#> 224 0.58012168
#> 225 0.58142471
#> 226 0.58398799
#> 227 0.58643438
#> 228 0.58963980
#> 229 0.59490664
#> 230 0.59954419
#> 231 0.60276047
#> 232 0.60495338
#> 233 0.60410206
#> 234 0.60712263
#> 235 0.61260616
#> 236 0.61963031
#> 237 0.62156355
#> 238 0.62297233
#> 239 0.62330147
#> 240 0.62731448
#> 241 0.62998947
#> 242 0.63216751
#> 243 0.63711831
#> 244 0.64447676
#> 245 0.64698117
#> 246 0.64452479
#> 247 0.63629304
#> 248 0.64538176
#> 249 0.65282532
#> 250 0.65827965
#> 251 0.66109927
#> 252 0.66061982
#> 253 0.66085028
#> 254 0.66523105
#> 255 0.66427269
#> 256 0.66487539
#> 257 0.67063301
#> 258 0.68098785
#> 259 0.69114508
#> 260 0.70309653
#> 261 0.70542253
#> 262 0.71225257
#> 263 0.71754015
#> 264 0.72456909
#> 265 0.73148166
#> 266 0.72916081
#> 267 0.72392156
#> 268 0.72485048
#> 269 0.72717139
#> 270 0.72203943
#> 271 0.71913396
#> 272 0.71474226
#> 273 0.71154515
#> 274 0.71212522
#> 275 0.71396122
#> 276 0.71293653
#> 277 0.72338923
#> 278 0.71745068
#> 279 0.72212899
#> 280 0.71518121
#> 281 0.70787528
#> 282 0.70829267
#> 283 0.71034890
#> 284 0.72379676
#> 285 0.71960200
#> 286 0.73282754
#> 287 0.72763234
#> 288 0.73381267
#> 289 0.72149528
#> 290 0.72805848
#> 291 0.72907084
#> 292 0.73564926
#> 293 0.73420160
#> 294 0.75257165
#> 295 0.76127330
#> 296 0.78237710
#> 297 0.76637290
#> 298 0.78792503
#> 299 0.78441192
#> 300 0.77311978
#> 301 0.77842377
#> 302 0.88278621
#> 303 0.92852079
#> 304 0.84178014
#> 305 0.73834894
#> 306 0.73334361
#> 307 0.70157061
#> 308 0.74278797
#> 309 0.76365512
#> 310 0.79062567
#> 311 0.79768489
#> 312 0.77769146
#> 313 0.74252517
#> 314 0.77052664
#> 315 0.76049144
#> 316 0.75444854
#> 317 0.74161152
#> 318 0.75426979
#> 319 0.75495611
#> 320 0.75472174
#> 321 0.76205897
#> 322 0.76148239
#> 323 0.76947565
#> 324 0.77236616
#> 325 0.76422096
#> 326 0.76138973
#> 327 0.76083280
#> 328 0.76146042
#> 329 0.76729635
#> 330 0.77164141
#> 331 0.77372230
#> 332 0.77001535
#> 333 0.76263056
#> 334 0.76035928
#> 335 0.76100106
#> 336 0.76488550
#> 337 0.76501398
#> 338 0.76275310
#> 339 0.75592164
#> 340 0.75637255
#> 341 0.75379579
#> 342 0.74920875
#> 343 0.74720857
#> 344 0.75329105
#> 345 0.74165817
#> 346 0.72442665
#> 347 0.69684112
#> 348 0.69809589
#> 349 0.69777685
#> 350 0.69381105
#> 351 0.68728762
#> 352 0.67250729
#> 353 0.66810550
#> 354 0.66355772
#> 355 0.65148562
#> 356 0.64408642
#> 357 0.64169442
#> 358 0.63558045
#> 359 0.63208175
#> 360 0.63238194
#> 361 0.62653864
#> 362 0.62550452
#> 363 0.62391714
#> 364 0.62089282
#> 365 0.62028515
#> 366 0.61657130
#> 367 0.61041119
#> 368 0.60717563
#> 369 0.60378194
#> 370 0.59980116
#> 371 0.59644083
#> 372 0.59333494
#> 373 0.59006751
#> 374 0.58723505
#> 375 0.58520583
#> 376 0.58355819
#> 377 0.58326758
#> 378 0.57793888
#> 379 0.57833852
#> 380 0.57483396
#> 381 0.57137108
#> 382 0.56892827
#> 383 0.56720619
#> 384 0.56636411
#> 385 0.56212203
#> 386 0.56222888
#> 387 0.55959158
#> 388 0.55795132
#> 389 0.55286058
#> 390 0.55133273
#> 391 0.54998678
#> 392 0.54983752
#> 393 0.54793034
#> 394 0.54823340
#> 395 0.54623549
#> 396 0.54597722
#> 397 0.54076738
#> 398 0.54061106
#> 399 0.53833004
#> 400 0.53497661
#> 401 0.53383937
#> 402 0.53417858
#> 403 0.53306249